Remove cargo2android.json.
cargo2android.py is now deprecated in favour of cargo_embargo. See
https://android.googlesource.com/platform/development/+/main/tools/cargo_embargo/README.md
for details on how to use it. I have been going through converting
existing cargo2android configs to cargo_embargo, but as the nearby
cargo2android config has been broken for a while I'm deleting it for
now instead, as discussed.

Test: N/A
Change-Id: I8bc2505f5d596b1031d1c0bab923960da9ea7bc9