Remove unused Kotlin source glob from srcs
Putting an entry in srcs that ends with ".kt" causes Soong to add
a dependency on the Kotlin standard library.  In this case, there
are no sources that match the glob, so the dependency ends up
being ignored.  A change in Soong will cause it to honor the
dependency, which results in the Kotlin standard library ending
up in services.jar.  Remove the unused Kotlin source glob here.

Bug: 308016794
Test: m services
Flag: EXEMPT refactor
Change-Id: Ia5ebaf8316883d447027d9e685eb14107aab9987